Base de datos TERMINADA.
FUNCIONAN TODAS MIS RELACIONES JAJAJAJAJAJA BIEN POR FIIIIIN.

+++ database/migrations/2019_06_06_213945_create_votos_table.php
@@ -15,7 +15,7 @@ class CreateVotosTable extends Migration
     {
       Schema::create('votos', function (Blueprint $table) {
           $table->bigIncrements('id');
-          $table->integer('imagenes_id')->unsigned();
+          $table->bigInteger('imagenes_id')->unsigned();
           $table->foreign('imagenes_id')->references('id')->on('imagenes')->onDelete('cascade');
           $table->timestamps();
       });
